ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.4",
    "jupyterlab>=4.3.0",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.5",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "requests>=2.32.3",
    "uvicorn>=0.32.0",
    "pymupdf>=1.22.3",
    "fitz>=0.0.1.dev2",
]
